correctly quote identifiers involving reserved keywords (#22)
This PR directly addresses a real-world issue I discovered where dumping
the schema of a database that had a column named `offset` resulted in
SQL that could not be applied: the column name `offset` requires
escaping (`"offset"`) in order to be legal in a `CREATE TABLE`
statement.

It turns out there are a large number of such identifiers, [all
assiduously documented by
Postgres](https://www.postgresql.org/docs/current/sql-keywords-appendix.html),
which require quoting in some or all contexts where they might be used:

> In the PostgreSQL parser, life is a bit more complicated. There are
several different classes of tokens ranging from those that can never be
used as an identifier to those that have absolutely no special status in
the parser, but are considered ordinary identifiers. (The latter is
usually the case for functions specified by SQL.) Even reserved key
words are not completely reserved in PostgreSQL, but can be used as
column labels (for example, SELECT 55 AS CHECK, even though CHECK is a
reserved key word).
>
> [...]
>
> As a general rule, if you get spurious parser errors for commands that
use any of the listed key words as an identifier, you should try quoting
the identifier to see if the problem goes away.

This PR:

- Moves all quoting/escaping logic to the `pgtools.Literal` (for quoting
literals) and `pgtools.Identifier` (for quoting identifier) functions.
Previously, escaping was implemented there as well as in some unexported
helper functions within the `schema` package.
- Updates the `pgtools.Identifier` implementation to explicitly quote
any keywords that require quoting in at least _some_ contexts, which are
those from the Postgres documentation that are marked:
    - reserved
    - non-reserved (cannot be function or type)
    - reserved (can be function or type)
    - requires AS
- Simplifies/clarifies/improves the tests for the `pgtools.Literal` and
`pgtools.Identifier` functions.
- Adds a regression test to the `schema` package to confirm that this PR
fixes the real-world issue.

## Testing
- Existing test coverage confirms that the standardization on `pgtools`
for quoting/escaping is safe and correct.
- New unit tests in the `pgtools` package for the quoting logic
- New db-backed test in the `schema` package confirms the fix for
`pgmigrate dump`