fix: internal WithDB helper didnt work with postgres:17
The tests were failing repeatedly because the very long test database
names that were being generated by WithDB were getting truncated to 63
characters. When the helper tried to connect to the database with its
full >63-character name, it wouldn't be able to connect because that
full name didn't exist.

This commit fixes the issue by adopting the shorter, just-as-random
names generated by the logic introduced in
https://github.com/peterldowns/pgtestdb/pull/17

I tested by running `just test` and saw that everything passed.