fix: various issues with postgres:17 (#24)
After upgrading the docker-compose.yml for local tests to postgres:17, I
observed a variety of failures. This PR fixes those failures, and
updates the CI setup to also use postgres:17.

Issues discovered in 17:

- The tests were failing repeatedly because the very long test database
names that were being generated by WithDB were getting truncated to 63
characters. When the helper tried to connect to the database with its
full >63-character name, it wouldn't be able to connect because that
full name didn't exist. The fix is to update the WithDB helper by
adopting the shorter, just-as-random names generated by the logic
introduced in https://github.com/peterldowns/pgtestdb/pull/17.
- The definition of materialized views seems to no longer include the
name of the table when it isn't explicitly used as an alias, so I had to
update one of the tests.
- There was an issue in postgres 17 (fixed in 17.3) where domains with
NOT NULL constraints couldn't be parsed correctly due to
`pg_get_constraintdef` failing on the not-null constraint. This was
fixed upstream in 17.3 and I've also added a workaround in the code that
fixed the problem when I ran on 17.0.

## Testing
- [x] locally running `just test` and seeing the previously-failing
tests pass.
- [x] update CI to use postgres:17 and seeing it pass.