feat: add ls as alias for list (#27)
This PR makes it possible to run `pgmigrate ls` as an alias for
`pgmigrate list`, which is already an alias for `pgmigrate applied`.
Solves a long-standing pet peeve of mine, I think all CLIs should let
`ls` == `list`.

## Tests
I manually built and ran `pgmigrate ls` and `pgmigrate list` and saw
that they did the same thing. We don't have any tests on the CLI.